requirements[1].txt
flask
azure-storage-blob
azure-identity
azure-keyvault-secrets
python-dotenv